+++ magda-web-client/src/Components/SearchFacets/FacetBasicBody.js
@@ -1,9 +1,9 @@
 import React, { Component } from "react";
-import find from "lodash.find";
-import maxBy from "lodash.maxby";
+import find from "lodash/find";
+import maxBy from "lodash/maxBy";
 import defined from "../../helpers/defined";
 import FacetSearchBox from "./FacetSearchBox";
-import "./FacetBasicBody.css";
+import "./FacetBasicBody.scss";
 import { NoResultsLabel } from "./NoResultsLabel";
 
 class FacetBasicBody extends Component {