+++ CHANGES.md
@@ -1,10 +1,14 @@
 ## 0.0.54
 
+UI:
+
+-   Upgraded the to create-react-app 2
+-   Added a basic new dataset page
+
 Others:
 
 -   Change content api delete response status code from to 204 to 200 so its content can be read.
 -   Added value to the "Search result and click" Google Analytics event that will show averages and other numeric aggregates in Google Analytics.
-
 
 ## 0.0.53